services/legal_ai/requirements.txt
fastapi
uvicorn
motor
pypdf
python-multipart
python-dotenv
google-generativeai